Set manual GT level to float

+++ fms_fsdp/utils/dataloader_utils.py
@@ -34,7 +34,7 @@ def causal_lm(data_seq):
     data_seq = data_seq[:-1]
     diff_seq = t.view(-1,128)
     resample_interval = 0 #torch.rand(diff_seq.size(0),1).sqrt()
-    resample_interval = torch.ones_like(diff_seq) * resample_interval
+    resample_interval = torch.ones_like(diff_seq).float() * resample_interval
     resample_mask = torch.bernoulli(resample_interval).int()
     prev_shuffle = diff_seq.roll(1, dims=0)
     prev_shuffle = torch.stack([x[0,torch.randperm(128)] for x in prev_shuffle.split(1)], dim=0)
